move tests to tests/
